Add OTA function Enable static IP option in easylink

+++ EMW ToolBox/CustomIOS7AlertView/CustomIOS7AlertView.h
@@ -28,7 +28,7 @@
 @property (nonatomic, retain) NSArray *buttonTitles;
 @property (nonatomic, assign) BOOL useMotionEffects;
 
-@property (copy) void (^onButtonTouchUpInside)(CustomIOS7AlertView *alertView, int buttonIndex) ;
+@property (copy) void (^onButtonTouchUpInside)(CustomIOS7AlertView *alertView, NSInteger buttonIndex) ;
 
 - (id)init;
